requirements-freeze.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.0
Deprecated==1.2.18
Django==5.0.7
django-environ==0.11.2
filelock==3.16.1
flake8==7.1.0
fsspec==2024.9.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
huggingface-hub==0.25.1
idna==3.10
importlib-metadata==7.0.0
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.5.3
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opentelemetry-api==1.33.0
opentelemetry-exporter-otlp==1.33.0
opentelemetry-exporter-otlp-proto-common==1.33.0
opentelemetry-exporter-otlp-proto-grpc==1.33.0
opentelemetry-exporter-otlp-proto-http==1.33.0
opentelemetry-instrumentation==0.54b0
opentelemetry-instrumentation-celery==0.54b0
opentelemetry-instrumentation-django==0.54b0
opentelemetry-instrumentation-logging==0.54b0
opentelemetry-instrumentation-wsgi==0.54b0
opentelemetry-proto==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
opentelemetry-util-http==0.54b0
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==5.29.4
psycopg2==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.1
pytest-django==4.8.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.1
sympy==1.13.3
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.43.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.21.0
